tons of changes to make everything compile
mostly tons of small, distributed changes, primarily to remove all usage of `undelegateIfPossible` parameters

+++ src/test/utils/Operators.sol
@@ -66,8 +66,8 @@ contract Operators is Test {
 
     function stringToUint(string memory s) public pure returns (uint) {
         bytes memory b = bytes(s);
-        uint result = 0;
-        for (uint i = 0; i < b.length; i++) {
+        uint256 result = 0;
+        for (uint256 i = 0; i < b.length; i++) {
             if (uint256(uint8(b[i])) >= 48 && uint256(uint8(b[i])) <= 57) {
                 result = result * 10 + (uint256(uint8(b[i])) - 48); 
             }
